fix(driver-app): surface Android Auto failures to Crashlytics; isolate degrade test
Address confirmed code-review findings on the native-require guard:

- Both failure paths in registerAutoPlay (missing native module, listener
  registration failure) now call recordNonFatal() in addition to a local
  console.error. console output reaches no dashboard (Sentry drops console
  breadcrumbs; nothing wires console to Crashlytics/LogRocket), and the old
  bottom catch was dev-only log() — fully silent in production. A cohort
  losing car support is now remotely visible with searchable attributes.
- Move the missing-module test to register.degrade.test.ts with a hoisted
  throwing factory: the doMock/dontMock surgery in the shared file left the
  registry without its baseline iternio mock for any test appended later
  (order-dependent breakage), and its toHaveLength(0) assertion was
  non-discriminating. The new test asserts the Crashlytics non-fatal too.
- Correct the docs/app.config record: the runtimeVersion 2.3.0 bump landed
  in the same merge as the iternio/nitro deps, so OTA to a pre-iternio
  binary was already fenced — the earlier doc claim that it was never
  bumped was wrong; the guard is Expo Go coverage + belt-and-suspenders.

57/57 androidAuto tests pass (6 suites); lint + tsc clean on touched files.

Co-developed with Claude Code (claude-sonnet-4-6)
