+++ driver-app/lib/androidAuto/__tests__/register.test.ts
@@ -280,32 +280,8 @@ it('never creates a template when no car is connected', () => {
   expect(mockTemplates).toHaveLength(0);
 });
 
-it('degrades to no-car-support instead of throwing when the native module is missing', () => {
-  // Simulates Expo Go / an older binary receiving this JS via OTA: importing
-  // @iternio/react-native-auto-play instantiates its Nitro HybridObject, which
-  // throws when the native side is absent. registerAutoPlay runs at bundle
-  // load (index.js), so a throw here would crash the whole phone app.
-  const errSpy = jest.spyOn(console, 'error').mockImplementation(() => {});
-  jest.resetModules();
-  jest.doMock('@iternio/react-native-auto-play', () => {
-    throw new Error('Nitro HybridObject "AutoPlay" is not registered');
-  });
-  try {
-    // The reset registry re-evaluates react-native with the preset default OS
-    // (ios) — force android or the guard exits before requiring the package.
-    // eslint-disable-next-line @typescript-eslint/no-require-imports
-    (require('react-native').Platform as { OS: string }).OS = 'android';
-    // eslint-disable-next-line @typescript-eslint/no-require-imports
-    const register = require('../register').default as typeof registerAutoPlay;
-    expect(() => register()).not.toThrow();
-    expect(errSpy).toHaveBeenCalledWith(
-      expect.stringContaining('native module unavailable'),
-      expect.any(Error),
-    );
-    expect(mockTemplates).toHaveLength(0);
-  } finally {
-    errSpy.mockRestore();
-    jest.dontMock('@iternio/react-native-auto-play');
-    jest.resetModules();
-  }
-});
+// The missing-native-module degrade path is covered in register.degrade.test.ts:
+// it needs a THROWING @iternio mock as the hoisted module factory, and doing
+// that in this file via doMock/dontMock surgery leaves the module registry
+// without its baseline mock for any test appended later (order-dependent
+// breakage) — so it lives in its own file with its own registry.
